requirements.txt
asyncio==3.4.3
aiofiles==23.2.1
aiohttp==3.9.5
numpy==1.26.4
pillow==10.2.0
psutil
git+https://github.com/KurimuzonAkuma/pyrogram
py-tgcalls==0.9.7
python-dotenv==1.0.1
tgcrypto==1.2.5
speedtest-cli==2.1.3
git+https://github.com/yt-dlp/yt-dlp@master
git+https://github.com/alexmercerind/youtube-search-python@main
git+https://github.com/joetats/youtube_search@master
